move pagination to script
